move SetMap into src/vs/base/common/map.ts
