+++ src/vs/editor/contrib/inlineCompletions/browser/provideInlineCompletions.ts
@@ -6,7 +6,7 @@
 import { assertNever } from 'vs/base/common/assert';
 import { DeferredPromise } from 'vs/base/common/async';
 import { CancellationToken } from 'vs/base/common/cancellation';
-import { SetMap } from 'vs/base/common/collections';
+import { SetMap } from 'vs/base/common/map';
 import { onUnexpectedExternalError } from 'vs/base/common/errors';
 import { IDisposable } from 'vs/base/common/lifecycle';
 import { ISingleEditOperation } from 'vs/editor/common/core/editOperation';
